Clean up sidebar component
